requirements.txt
asttokens==2.4.1
blinker==1.8.2
Cartopy==0.23.0
certifi==2024.8.30
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==2.1.0
Flask==3.0.3
fonttools==4.53.1
idna==3.10
importlib_metadata==8.5.0
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
netCDF4==1.7.1.post2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
plotly==5.24.1
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
tenacity==9.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
xarray==2024.9.0
zipp==3.20.2
